nice to look at fe, + rev comments
